Incorporating review comments: renamed packages for samples and renamed maven plugin module
