+++ samples/groovy/src/test/groovy/com/smartbear/readyapi4j/samples/groovy/RestRequestTest.groovy
@@ -1,4 +1,4 @@
-package com.smartbear.readyapi4j
+package com.smartbear.readyapi4j.samples.groovy
 
 import com.smartbear.readyapi4j.execution.Execution
 import com.smartbear.readyapi4j.execution.RecipeExecutor
@@ -9,70 +9,63 @@ import static com.smartbear.readyapi4j.dsl.execution.RecipeExecution.executeReci
 import static com.smartbear.readyapi4j.dsl.execution.RecipeExecution.executeRecipeOnServer
 import static com.smartbear.readyapi4j.dsl.execution.RecipeExecution.remoteRecipeExecutor
 
-class SoapRequestTest {
+class RestRequestTest {
+
     private static final String TEST_SERVER_URL = 'http://testserver.readyapi.io:8080'
     private static final String TEST_SERVER_USER = 'demoUser'
     private static final String TEST_SERVER_PASSWORD = 'demoPassword'
 
     @Test
-    void simpleSoapRequestUsingSoapUIEngine() throws Exception {
+    void testSimpleCountUsingSoapUIEngine() throws Exception {
+        //This requires additional dependency com.smartbear.readyapi:readyapi4j-local
         Execution execution = executeRecipe {
-            soapRequest {
-                wsdl = 'http://www.webservicex.com/globalweather.asmx?WSDL'
-                binding = 'GlobalWeatherSoap12'
-                operation = 'GetWeather'
-                namedParam 'CountryName', 'Sweden'
-                pathParam '//*:CityName', 'Stockholm'
+            get 'https://api.swaggerhub.com/apis', {
+                parameters {
+                    query 'query', 'testserver'
+                }
                 asserting {
-                    notSoapFault
-                    schemaCompliance
+                    status 200
+                    jsonPath '$.totalCount' occurs 1 times
                 }
             }
         }
-
         assert execution.currentStatus == FINISHED
         assert execution.errorMessages.empty
     }
 
     @Test
-    void simpleSoapRequestRunUsingTestServer() throws Exception {
+    void testSimpleCountByRunningRecipeOnTestServer() throws Exception {
         Execution execution = executeRecipeOnServer TEST_SERVER_URL, TEST_SERVER_USER, TEST_SERVER_PASSWORD, {
-            soapRequest {
-                wsdl = 'http://www.webservicex.com/globalweather.asmx?WSDL'
-                binding = 'GlobalWeatherSoap12'
-                operation = 'GetWeather'
-                namedParam 'CountryName', 'Sweden'
-                pathParam '//*:CityName', 'Stockholm'
+            get 'https://api.swaggerhub.com/apis', {
+                parameters {
+                    query 'query', 'testserver'
+                }
                 asserting {
-                    notSoapFault
-                    schemaCompliance
+                    status 200
+                    jsonPath '$.totalCount' occurs 1 times
                 }
             }
         }
-
         assert execution.currentStatus == FINISHED
         assert execution.errorMessages.empty
     }
 
     @Test
-    void simpleSoapRequestRunUsingTheProvidedExecutor() throws Exception {
+    void testSimpleCountByRunningWithTheProvidedExecutor() throws Exception {
         //The executor can be initialized in set-up method if used in multiple tests
         RecipeExecutor executor = remoteRecipeExecutor(TEST_SERVER_URL, TEST_SERVER_USER, TEST_SERVER_PASSWORD)
 
         Execution execution = executeRecipe executor, {
-            soapRequest {
-                wsdl = 'http://www.webservicex.com/globalweather.asmx?WSDL'
-                binding = 'GlobalWeatherSoap12'
-                operation = 'GetWeather'
-                namedParam 'CountryName', 'Sweden'
-                pathParam '//*:CityName', 'Stockholm'
+            get 'https://api.swaggerhub.com/apis', {
+                parameters {
+                    query 'query', 'testserver'
+                }
                 asserting {
-                    notSoapFault
-                    schemaCompliance
+                    status 200
+                    jsonPath '$.totalCount' occurs 1 times
                 }
             }
         }
-
         assert execution.currentStatus == FINISHED
         assert execution.errorMessages.empty
     }
