+++ samples/maven-plugin/pom.xml
