+++ samples/maven-plugin/src/test/resources/recipes/simple-soap-request.json
