+++ samples/maven-plugin/src/test/resources/recipes/simple.json
